+++ examples/super_stream/producer/super_stream_producer.go
@@ -98,7 +98,7 @@ func main() {
 				}
 			}
 		}
-	}(superStreamProducer.NotifyPublishConfirmation(1))
+	}(superStreamProducer.NotifyPublishConfirmation(0))
 
 	// Publish messages
 loop: